requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
-e git+https://github.com/news-polygraph/bachelor-thesis-eltayeb.git@7999a6b30f613a76ddca7e18b499cd2252d78af4#egg=BACHELOR_THESIS_ELTAYEB
beir==2.2.0
bitarray==3.4.2
blinker==1.9.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
colbert-ai==0.2.19
coloredlogs==15.0.1
confection==0.1.5
cymem==2.0.11
Cython==3.1.2
dataclasses-json==0.6.7
datasets==3.6.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
et_xmlfile==2.0.0
faiss-cpu==1.11.0
filelock==3.18.0
Flask==3.1.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
hf-xet==1.1.4
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-openai==0.3.23
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.45
language_data==1.3.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.4.4
multiprocess==0.70.16
murmurhash==1.0.13
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
ninja==1.11.1.4
numpy==2.3.0
omegaconf==2.3.0
onnxruntime==1.22.0
openai==1.88.0
openpyxl==3.1.5
orjson==3.10.18
packaging==24.2
pandas==2.3.0
pillow==11.2.1
preshed==3.0.10
propcache==0.3.2
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyjnius==1.6.1
pyserini==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytrec_eval-terrier==0.5.7
pytz==2025.2
PyYAML==6.0.2
-e git+https://github.com/Abdalla1423/RAG_comparison.git@42cba35365a57edf6342c0bbb71ddec8227d67de#egg=rag_benchmark
ragas==0.2.15
rank-bm25==0.2.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.0.0
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.19.1
torch==2.7.1
tqdm==4.67.1
transformers==4.41.2
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ujson==5.10.0
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
